refactor(server): fold role-repository into role-queries
Delete the 23-line shallow repositories/role-repository.ts (one function,
one consumer) and move getRoleInUseCount into db/role-queries.ts next to
its six siblings. routes/roles.ts now imports it from there, and
roles.test.ts's mock + dynamic imports follow.

With both half-applied seams gone, data access converges on one
convention repo-wide: db/<entity>-queries.ts holds the SQL, and the
refresh-token module is the single deep repository.

Refs #1235
